backend/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
oauthlib==3.2.2
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
fastapi
uvicorn[standard]